requirements.txt
pillow
streamlit
numpy
pandas
tensorflow